metrics: Concatenate go files under metrics dir
